shitpost.js Module, Travis Test Changes, Package.json update
