+++ README.md
@@ -8,6 +8,7 @@ Bot utilizing [Discord.js](https://github.com/hydrabolt/discord.js) as it's prim
 - !configtest command for testing your config
 - !fish command to slap yourself about with one of 1,068 species of fish
 - !roulette to select a user at random that is online
+- !shitpost command to place shitpost from certain subreddits
 - Config.json controls what functions are enabled in the bot, by default it is currently everything
 
 ##Planned Features for 1.0##
