+++ package.json
@@ -5,14 +5,16 @@
   "version": "0.4.0",
   "description": "Bot for Discord based on discord.js",
   "repository": {
-    "type" : "git",
-	"url": "https://github.com/dricephd/DrIceDiscordBot"
+    "type": "git",
+    "url": "https://github.com/dricephd/DrIceDiscordBot"
   },
   "main": "client.js",
   "dependencies": {
-    "discord.js": "^5.1.0",
+    "discord.js": "5.2.0",
     "node.js": "0.0.0",
-    "request": "^2.67.0"
+    "request": "2.67.0"
   },
-  "scripts": { "test": "node test/req-test.js" }
+  "scripts": {
+    "test": "node test/travistest.js"
+  }
 }
